backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
aiohttp
beautifulsoup4
pypdf
chromadb>=0.4.18
sentence-transformers
torch>=2.2.0
playwright
python-multipart
openai
textual